Revert "update to 7.5 (Incomplete and untested)"
This reverts commit cea68f81e52a6f268370814844cee79fe7e9d60c.

+++ others/py3compat.py
@@ -1,7 +1,7 @@
 
 # Workaround for Python3 compat
 try:
-  INTEGER_TYPES = (int, int)
+  INTEGER_TYPES = (int, long)
 except NameError:
   long = int
   INTEGER_TYPES = (int,)
